# 비트 카운팅 정상동작 확인

+++ Code/Majestro/Majestro_Engine/BeatSystem.h
@@ -14,7 +14,7 @@ public:
 	void Update(float dt);
 
 private:
-	int mBpm = 120;
+	float mBpm = 120.0f;
 	int mBeat =0;
 	float mSeconds = 0.0f;
 };
